Renamed test file appropriately
